requirements.txt
pypresence>=4.3.0
requests>=2.28.0
pystray>=0.19.0
Pillow>=9.0.0
python-dotenv>=1.0.0
